refactor: migrate all presenters to Metro and AssistedInject
Refactors the presentation layer to replace manual factory implementations and `kotlin-inject` assisted injection with Metro's `@AssistedInject` and `@AssistedFactory`.

Changes:
- Replaced `Default*Presenter` implementations with single `@AssistedInject` classes.
- Replaced manual `Factory` interfaces with Metro `@AssistedFactory` functional interfaces.
- Updated all presenter factories to use the generated `create` method instead of `invoke`.
- Migrated internal `PresenterInstance` (used in Discover and Search) to use Metro-managed dependencies.
- Standardized `componentContext.coroutineScope()` usage across all presenters.
- Updated related unit tests to reflect the removal of `Default*` prefixed implementation classes.
- Removed all `Default*Presenter.kt` files and consolidated logic into the primary presenter classes.
