--- presenter/discover/src/commonMain/kotlin/com/thomaskioko/tvmaniac/discover/presenter/DefaultDiscoverShowsPresenter.kt
@@ -1,295 +1,0 @@
-package com.thomaskioko.tvmaniac.discover.presenter
-
-import com.arkivanov.decompose.ComponentContext
-import com.arkivanov.decompose.value.Value
-import com.arkivanov.essenty.instancekeeper.InstanceKeeper
-import com.arkivanov.essenty.instancekeeper.getOrCreate
-import com.thomaskioko.tvmaniac.core.base.annotations.ActivityScope
-import com.thomaskioko.tvmaniac.core.base.extensions.asValue
-import com.thomaskioko.tvmaniac.core.base.extensions.combine
-import com.thomaskioko.tvmaniac.core.base.extensions.coroutineScope
-import com.thomaskioko.tvmaniac.core.logger.Logger
-import com.thomaskioko.tvmaniac.core.view.ErrorToStringMapper
-import com.thomaskioko.tvmaniac.core.view.ObservableLoadingCounter
-import com.thomaskioko.tvmaniac.core.view.UiMessageManager
-import com.thomaskioko.tvmaniac.core.view.collectStatus
-import com.thomaskioko.tvmaniac.data.featuredshows.api.interactor.FeaturedShowsInteractor
-import com.thomaskioko.tvmaniac.data.popularshows.api.PopularShowsInteractor
-import com.thomaskioko.tvmaniac.data.upcomingshows.api.UpcomingShowsInteractor
-import com.thomaskioko.tvmaniac.discover.api.TrendingShowsInteractor
-import com.thomaskioko.tvmaniac.discover.presenter.model.NextEpisodeUiModel
-import com.thomaskioko.tvmaniac.domain.discover.DiscoverShowsInteractor
-import com.thomaskioko.tvmaniac.domain.episode.MarkEpisodeWatchedInteractor
-import com.thomaskioko.tvmaniac.domain.episode.MarkEpisodeWatchedParams
-import com.thomaskioko.tvmaniac.domain.followedshows.UnfollowShowInteractor
-import com.thomaskioko.tvmaniac.domain.genre.GenreShowsInteractor
-import com.thomaskioko.tvmaniac.followedshows.api.FollowedShowsRepository
-import com.thomaskioko.tvmaniac.shows.api.model.Category
-import com.thomaskioko.tvmaniac.topratedshows.data.api.TopRatedShowsInteractor
-import com.thomaskioko.tvmaniac.traktauth.api.TraktAuthRepository
-import com.thomaskioko.tvmaniac.traktauth.api.TraktAuthState
-import com.thomaskioko.tvmaniac.upnext.api.model.NextEpisodeWithShow
-import kotlinx.collections.immutable.toImmutableList
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.cancel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.drop
-import kotlinx.coroutines.flow.filter
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.launch
-import me.tatarka.inject.annotations.Assisted
-import me.tatarka.inject.annotations.Inject
-import software.amazon.lastmile.kotlin.inject.anvil.ContributesBinding
-
-@Inject
-@ContributesBinding(ActivityScope::class, DiscoverShowsPresenter::class)
-public class DefaultDiscoverShowsPresenter(
-    @Assisted componentContext: ComponentContext,
-    @Assisted private val onNavigateToShowDetails: (Long) -> Unit,
-    @Assisted private val onNavigateToMore: (Long) -> Unit,
-    @Assisted private val onNavigateToEpisode: (showTraktId: Long, episodeId: Long) -> Unit,
-    @Assisted private val onNavigateToSeason: (showTraktId: Long, seasonId: Long, seasonNumber: Long) -> Unit,
-    @Assisted private val onNavigateToUpNext: () -> Unit,
-    @Assisted private val onNavigateToSearch: () -> Unit,
-    private val discoverShowsInteractor: DiscoverShowsInteractor,
-    private val followedShowsRepository: FollowedShowsRepository,
-    private val unfollowShowInteractor: UnfollowShowInteractor,
-    private val featuredShowsInteractor: FeaturedShowsInteractor,
-    private val topRatedShowsInteractor: TopRatedShowsInteractor,
-    private val popularShowsInteractor: PopularShowsInteractor,
-    private val trendingShowsInteractor: TrendingShowsInteractor,
-    private val upcomingShowsInteractor: UpcomingShowsInteractor,
-    private val genreShowsInteractor: GenreShowsInteractor,
-    private val markEpisodeWatchedInteractor: MarkEpisodeWatchedInteractor,
-    private val traktAuthRepository: TraktAuthRepository,
-    private val errorToStringMapper: ErrorToStringMapper,
-    private val logger: Logger,
-    private val coroutineScope: CoroutineScope = componentContext.coroutineScope(),
-) : DiscoverShowsPresenter, ComponentContext by componentContext {
-
-    override val presenterInstance: PresenterInstance = instanceKeeper.getOrCreate { PresenterInstance() }
-
-    override val state: StateFlow<DiscoverViewState> = presenterInstance.state
-    override val stateValue: Value<DiscoverViewState> = state.asValue(coroutineScope)
-
-    init {
-        presenterInstance.init()
-    }
-
-    override fun dispatch(action: DiscoverShowAction) {
-        presenterInstance.dispatch(action)
-    }
-
-    public inner class PresenterInstance : InstanceKeeper.Instance {
-
-        private val featuredLoadingState = ObservableLoadingCounter()
-        private val topRatedLoadingState = ObservableLoadingCounter()
-        private val popularLoadingState = ObservableLoadingCounter()
-        private val trendingLoadingState = ObservableLoadingCounter()
-        private val upcomingLoadingState = ObservableLoadingCounter()
-        private val genreState = ObservableLoadingCounter()
-        private val upNextActionLoadingState = ObservableLoadingCounter()
-        private val uiMessageManager = UiMessageManager()
-
-        private val _state: MutableStateFlow<DiscoverViewState> = MutableStateFlow(
-            DiscoverViewState.Empty.copy(featuredRefreshing = true),
-        )
-        public val state: StateFlow<DiscoverViewState> = combine(
-            upNextActionLoadingState.observable,
-            featuredLoadingState.observable,
-            topRatedLoadingState.observable,
-            popularLoadingState.observable,
-            trendingLoadingState.observable,
-            upcomingLoadingState.observable,
-            discoverShowsInteractor.flow,
-            uiMessageManager.message,
-            _state,
-        ) {
-                upNextUpdating, featuredShowsIsUpdating, topRatedShowsIsUpdating, popularShowsIsUpdating,
-                trendingShowsIsUpdating, upComingIsUpdating,
-                showData, message, currentState,
-            ->
-
-            currentState.copy(
-                message = message,
-                featuredRefreshing = featuredShowsIsUpdating,
-                topRatedRefreshing = topRatedShowsIsUpdating,
-                popularRefreshing = popularShowsIsUpdating,
-                trendingRefreshing = trendingShowsIsUpdating,
-                upcomingRefreshing = upComingIsUpdating,
-                upNextRefreshing = upNextUpdating,
-                featuredShows = showData.featuredShows.toShowList(),
-                topRatedShows = showData.topRatedShows.toShowList(),
-                popularShows = showData.popularShows.toShowList(),
-                trendingToday = showData.trendingShows.toShowList(),
-                upcomingShows = showData.upcomingShows.toShowList(),
-                nextEpisodes = showData.nextEpisodes
-                    .map { it.toUiModel() }
-                    .toImmutableList(),
-            )
-        }.stateIn(
-            scope = coroutineScope,
-            started = SharingStarted.WhileSubscribed(),
-            initialValue = _state.value,
-        )
-
-        public fun init() {
-            discoverShowsInteractor(Unit)
-            observeShowData()
-            observeAuthState()
-        }
-
-        private fun observeAuthState() {
-            coroutineScope.launch {
-                traktAuthRepository.state
-                    .drop(1)
-                    .distinctUntilChanged()
-                    .filter { it == TraktAuthState.LOGGED_IN }
-                    .collect { observeShowData(forceRefresh = true) }
-            }
-        }
-
-        public fun dispatch(action: DiscoverShowAction) {
-            when (action) {
-                is ShowClicked -> onNavigateToShowDetails(action.traktId)
-                PopularClicked -> onNavigateToMore(Category.POPULAR.id)
-                TopRatedClicked -> onNavigateToMore(Category.TOP_RATED.id)
-                TrendingClicked -> onNavigateToMore(Category.TRENDING_TODAY.id)
-                UpComingClicked -> onNavigateToMore(Category.UPCOMING.id)
-                UpNextMoreClicked -> onNavigateToUpNext()
-                RefreshData -> observeShowData(forceRefresh = true)
-                is UpdateShowInLibrary -> {
-                    coroutineScope.launch {
-                        if (action.inLibrary) {
-                            unfollowShowInteractor.executeSync(action.traktId)
-                        } else {
-                            followedShowsRepository.addFollowedShow(action.traktId)
-                        }
-                    }
-                }
-                is MessageShown -> {
-                    clearMessage(action.id)
-                }
-                is NextEpisodeClicked -> onNavigateToSeason(action.showTraktId, action.seasonId, action.seasonNumber)
-                is MarkNextEpisodeWatched -> {
-                    coroutineScope.launch {
-                        markEpisodeWatchedInteractor(
-                            MarkEpisodeWatchedParams(
-                                showTraktId = action.showTraktId,
-                                episodeId = action.episodeId,
-                                seasonNumber = action.seasonNumber,
-                                episodeNumber = action.episodeNumber,
-                            ),
-                        ).collectStatus(upNextActionLoadingState, logger, uiMessageManager, errorToStringMapper = errorToStringMapper)
-                    }
-                }
-                is UnfollowShowFromUpNext -> {
-                    coroutineScope.launch {
-                        unfollowShowInteractor.executeSync(action.showTraktId)
-                    }
-                }
-                is OpenSeasonFromUpNext -> {
-                    onNavigateToSeason(action.showTraktId, action.seasonId, action.seasonNumber)
-                }
-                is OpenShowFromUpNext -> onNavigateToShowDetails(action.showTraktId)
-                SearchIconClicked -> onNavigateToSearch()
-                is DiscoverEpisodeLongPressed -> onNavigateToEpisode(action.showTraktId, action.episodeId)
-            }
-        }
-
-        internal fun clearMessage(id: Long) {
-            coroutineScope.launch {
-                uiMessageManager.clearMessage(id)
-            }
-        }
-
-        private fun observeShowData(forceRefresh: Boolean = false) {
-            coroutineScope.launch {
-                genreShowsInteractor(forceRefresh)
-                    .collectStatus(genreState, logger, uiMessageManager, "Genres", errorToStringMapper)
-            }
-            coroutineScope.launch {
-                featuredShowsInteractor(forceRefresh)
-                    .collectStatus(featuredLoadingState, logger, uiMessageManager, "Featured Shows", errorToStringMapper)
-            }
-
-            coroutineScope.launch {
-                topRatedShowsInteractor(forceRefresh)
-                    .collectStatus(topRatedLoadingState, logger, uiMessageManager, "Top Rated Shows", errorToStringMapper)
-            }
-
-            coroutineScope.launch {
-                popularShowsInteractor(forceRefresh)
-                    .collectStatus(popularLoadingState, logger, uiMessageManager, "Popular Shows", errorToStringMapper)
-            }
-
-            coroutineScope.launch {
-                trendingShowsInteractor(forceRefresh)
-                    .collectStatus(trendingLoadingState, logger, uiMessageManager, "Trending Shows", errorToStringMapper)
-            }
-
-            coroutineScope.launch {
-                upcomingShowsInteractor(forceRefresh)
-                    .collectStatus(upcomingLoadingState, logger, uiMessageManager, "Upcoming Shows", errorToStringMapper)
-            }
-        }
-
-        override fun onDestroy() {
-            coroutineScope.cancel()
-        }
-    }
-}
-
-@Inject
-@ContributesBinding(ActivityScope::class, DiscoverShowsPresenter.Factory::class)
-public class DefaultDiscoverPresenterFactory(
-    private val presenter: (
-        componentContext: ComponentContext,
-        onNavigateToShowDetails: (id: Long) -> Unit,
-        onNavigateToMore: (categoryId: Long) -> Unit,
-        onNavigateToEpisode: (showTraktId: Long, episodeId: Long) -> Unit,
-        onNavigateToSeason: (showTraktId: Long, seasonId: Long, seasonNumber: Long) -> Unit,
-        onNavigateToUpNext: () -> Unit,
-        onNavigateToSearch: () -> Unit,
-    ) -> DiscoverShowsPresenter,
-) : DiscoverShowsPresenter.Factory {
-    override fun invoke(
-        componentContext: ComponentContext,
-        onNavigateToShowDetails: (id: Long) -> Unit,
-        onNavigateToMore: (categoryId: Long) -> Unit,
-        onNavigateToEpisode: (showTraktId: Long, episodeId: Long) -> Unit,
-        onNavigateToSeason: (showTraktId: Long, seasonId: Long, seasonNumber: Long) -> Unit,
-        onNavigateToUpNext: () -> Unit,
-        onNavigateToSearch: () -> Unit,
-    ): DiscoverShowsPresenter = presenter(
-        componentContext,
-        onNavigateToShowDetails,
-        onNavigateToMore,
-        onNavigateToEpisode,
-        onNavigateToSeason,
-        onNavigateToUpNext,
-        onNavigateToSearch,
-    )
-}
-
-private fun NextEpisodeWithShow.toUiModel(): NextEpisodeUiModel {
-    return NextEpisodeUiModel(
-        showTraktId = showTraktId,
-        showName = showName,
-        imageUrl = stillPath ?: showPoster,
-        episodeId = episodeId,
-        episodeTitle = episodeName ?: "",
-        episodeNumberFormatted = "S${seasonNumber}E$episodeNumber",
-        seasonId = seasonId,
-        seasonNumber = seasonNumber,
-        episodeNumber = episodeNumber,
-        runtime = runtime?.let { "$it min" },
-        overview = overview ?: "",
-        isNew = false,
-        rating = rating,
-        voteCount = voteCount,
-    )
-}
