+++ presenter/home/src/iosTest/kotlin/com/thomaskioko/tvmaniac/presenter/home/HomePresenterIosTest.kt
@@ -1,10 +1,13 @@
 package com.thomaskioko.tvmaniac.presenter.home
 
-import com.thomaskioko.tvmaniac.testing.di.TestIosComponent
+import com.thomaskioko.tvmaniac.testing.di.TestIosGraph
+import dev.zacsweers.metro.createGraphFactory
 
 internal class HomePresenterIosTest : HomePresenterTest() {
-    private val testComponent: TestIosComponent = TestIosComponent.create()
+    private val testGraph: TestIosGraph by lazy {
+        createGraphFactory<TestIosGraph.Factory>().create()
+    }
 
     override val homePresenterFactory: HomePresenter.Factory
-        get() = testComponent.homePresenterFactory
+        get() = testGraph.homePresenterFactory
 }
