+++ presenter/home/src/jvmTest/kotlin/com/thomaskioko/tvmaniac/presenter/home/HomePresenterJvmTest.kt
@@ -1,10 +1,11 @@
 package com.thomaskioko.tvmaniac.presenter.home
 
-import com.thomaskioko.tvmaniac.testing.di.TestJvmComponent
-import com.thomaskioko.tvmaniac.testing.di.create
+import com.thomaskioko.tvmaniac.testing.di.TestJvmGraph
+import dev.zacsweers.metro.createGraphFactory
 
 internal class HomePresenterJvmTest : HomePresenterTest() {
-    private val testComponent: TestJvmComponent = TestJvmComponent::class.create()
+    private val testComponent: TestJvmGraph =
+        createGraphFactory<TestJvmGraph.Factory>().create()
 
     override val homePresenterFactory: HomePresenter.Factory
         get() = testComponent.homePresenterFactory
