--- presenter/progress/src/commonMain/kotlin/com/thomaskioko/tvmaniac/presentation/progress/DefaultProgressPresenter.kt
@@ -1,84 +1,0 @@
-package com.thomaskioko.tvmaniac.presentation.progress
-
-import com.arkivanov.decompose.ComponentContext
-import com.arkivanov.decompose.childContext
-import com.arkivanov.decompose.value.Value
-import com.thomaskioko.tvmaniac.core.base.annotations.ActivityScope
-import com.thomaskioko.tvmaniac.core.base.extensions.asValue
-import com.thomaskioko.tvmaniac.core.base.extensions.coroutineScope
-import com.thomaskioko.tvmaniac.presentation.calendar.CalendarPresenter
-import com.thomaskioko.tvmaniac.presentation.upnext.UpNextPresenter
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.update
-import me.tatarka.inject.annotations.Assisted
-import me.tatarka.inject.annotations.Inject
-import software.amazon.lastmile.kotlin.inject.anvil.ContributesBinding
-import software.amazon.lastmile.kotlin.inject.anvil.SingleIn
-
-@Inject
-@ContributesBinding(ActivityScope::class, ProgressPresenter::class)
-public class DefaultProgressPresenter(
-    @Assisted componentContext: ComponentContext,
-    @Assisted private val navigateToShowDetails: (showId: Long) -> Unit,
-    @Assisted private val navigateToSeasonDetails: (showTraktId: Long, seasonId: Long, seasonNumber: Long) -> Unit,
-    @Assisted private val onUpNextEpisodeLongPressed: (Long) -> Unit,
-    @Assisted private val onCalendarEpisodeLongPressed: (Long) -> Unit,
-    private val upNextPresenterFactory: UpNextPresenter.Factory,
-    private val calendarPresenterFactory: CalendarPresenter.Factory,
-) : ProgressPresenter, ComponentContext by componentContext {
-    private val coroutineScope = coroutineScope()
-    private val _state = MutableStateFlow(ProgressState())
-
-    override val state: StateFlow<ProgressState> = _state.asStateFlow()
-    override val stateValue: Value<ProgressState> = state.asValue(coroutineScope)
-
-    override val upNextPresenter: UpNextPresenter = upNextPresenterFactory(
-        componentContext = childContext(key = "UpNext"),
-        navigateToShowDetails = navigateToShowDetails,
-        navigateToSeasonDetails = navigateToSeasonDetails,
-        onEpisodeLongPressed = onUpNextEpisodeLongPressed,
-    )
-
-    override val calendarPresenter: CalendarPresenter = calendarPresenterFactory(
-        componentContext = childContext(key = "Calendar"),
-        navigateToShowDetails = navigateToShowDetails,
-        onEpisodeLongPressed = onCalendarEpisodeLongPressed,
-    )
-
-    override fun dispatch(action: ProgressAction) {
-        when (action) {
-            is ProgressAction.SelectPage -> {
-                _state.update { it.copy(selectedPage = action.index) }
-            }
-        }
-    }
-}
-
-@Inject
-@SingleIn(ActivityScope::class)
-@ContributesBinding(ActivityScope::class, ProgressPresenter.Factory::class)
-public class DefaultProgressPresenterFactory(
-    private val presenter: (
-        componentContext: ComponentContext,
-        navigateToShowDetails: (showId: Long) -> Unit,
-        navigateToSeasonDetails: (showTraktId: Long, seasonId: Long, seasonNumber: Long) -> Unit,
-        onUpNextEpisodeLongPressed: (Long) -> Unit,
-        onCalendarEpisodeLongPressed: (Long) -> Unit,
-    ) -> ProgressPresenter,
-) : ProgressPresenter.Factory {
-    override fun invoke(
-        componentContext: ComponentContext,
-        navigateToShowDetails: (showId: Long) -> Unit,
-        navigateToSeasonDetails: (showTraktId: Long, seasonId: Long, seasonNumber: Long) -> Unit,
-        onUpNextEpisodeLongPressed: (episodeId: Long) -> Unit,
-        onCalendarEpisodeLongPressed: (episodeId: Long) -> Unit,
-    ): ProgressPresenter = presenter(
-        componentContext,
-        navigateToShowDetails,
-        navigateToSeasonDetails,
-        onUpNextEpisodeLongPressed,
-        onCalendarEpisodeLongPressed,
-    )
-}
