+++ presenter/search/src/commonTest/kotlin/com/thomaskioko/tvmaniac/presenter/search/SearchShowsPresenterTest.kt
@@ -15,7 +15,6 @@ import com.thomaskioko.tvmaniac.i18n.StringResourceKey
 import com.thomaskioko.tvmaniac.i18n.testing.FakeLocalizer
 import com.thomaskioko.tvmaniac.search.presenter.CategoryChanged
 import com.thomaskioko.tvmaniac.search.presenter.ClearQuery
-import com.thomaskioko.tvmaniac.search.presenter.DefaultSearchShowsPresenter
 import com.thomaskioko.tvmaniac.search.presenter.Mapper
 import com.thomaskioko.tvmaniac.search.presenter.QueryChanged
 import com.thomaskioko.tvmaniac.search.presenter.SearchShowState
@@ -354,7 +353,7 @@ internal class SearchShowsPresenterTest {
 
     private fun buildPresenter(
         lifecycle: LifecycleRegistry = LifecycleRegistry(),
-    ): SearchShowsPresenter = DefaultSearchShowsPresenter(
+    ): SearchShowsPresenter = SearchShowsPresenter(
         componentContext = DefaultComponentContext(lifecycle = lifecycle),
         onNavigateToShowDetails = {},
         onNavigateToGenre = {},
